move notifications to components
