+++ app/utilities/layout.tsx
@@ -2,7 +2,7 @@ import Image from 'next/image';
 import Link from 'next/link';
 import React from 'react';
 import LeftNav from '@/app/components/leftNav';
-import Notifications from '@/app/views/notifications/notifications';
+import Notifications from '@/app/components/notifications';
 import { MicroLayout } from '@/app/views/quickstart-home/components/microLayout';
 
 export default function BenchmarkingLayout({
